Make timelord run a local dockerfile.
This is a bit easier to mannage instead of having to login and upload a new image to docker hun everytime we have to make changes.

Besides that we rip out harvester for the time being and depend upon a live instance for the development.
In the near future we want to have both projects in a monorepo! This is not sustainable.

We also added the watch command for the docker-compose command. This way we ensure that the development experience is reactive and does what you expect.